#14: created subcomponents for map
